+++ app/controllers/api/payments_controller.rb
@@ -1,27 +1,51 @@
 class Api::PaymentsController < Api::ApiController
 
 
-  skip_before_action :verify_authenticity_token
+  before_action :is_rider? , except: [:index, :show]
   
 
   def index
-    @payments = Payment.all
-    if @payments.empty?
-      render json: { message: "Payments not found" } , status: :no_content
+    # @payments = Payment.all
+    # if @payments.empty?
+    #   render json: { message: "Payments not found" } , status: :no_content
+    # else
+    #   render json: @payments , status: :ok
+    # end
+    @payments = current_user.userable.payments
+    if @payments.size == 0
+      render json: { message: "No Payments available for you "} , status: :no_content
     else
       render json: @payments , status: :ok
     end
+
   end
 
 
 
   def show
-    @payment = Payment.find_by(id: params[:id])
-    if @payment
-      render json: @payment, status: :ok
-    else
-      render json: { message: "Payment not found" } , status: :not_found
+    # @payment = Payment.find_by(id: params[:id])
+    # if @payment
+    #   render json: @payment, status: :ok
+    # else
+    #   render json: { message: "Payment not found" } , status: :not_found
+    # end
+
+    @payments = current_user.userable.payments
+    @payments_id_array = @payments.pluck(:id)
+
+    unless @payments_id_array.include?(params[:id].to_i)
+      render json: { message: "You are not authorized to view this page" } , status: :forbidden
+      return 
     end
+
+    @payment = Payment.find_by(id: params[:id].to_i)
+
+    if @payment
+      render json: @payment , status: :ok
+    else
+      render json: { message: "No payment available with the id #{params[:id].to_i}"} , status: :no_content
+    end
+
   end
 
   
@@ -29,7 +53,7 @@ class Api::PaymentsController < Api::ApiController
   def create
 
     @payment = Payment.new()
-    @payment.rider_id = params[:rider_id]
+    @payment.rider_id = current_user.userable.id
     @payment.driver_id = params[:driver_id]
     @payment.mode_of_payment = params[:mode_of_payment]
     @payment.amount = params[:amount]
@@ -48,7 +72,16 @@ class Api::PaymentsController < Api::ApiController
 
 
   def update
-    @payment = Payment.find(params[:id])
+    @payments = current_user.userable.payments
+    @payments_id_array = @payments.pluck(:id)
+
+    unless @payments_id_array.include?(params[:id].to_i)
+      render json: { message: "You are not authorized to view this page" } , status: :forbidden
+      return       
+    end
+
+    @payment = Payment.find_by(id: params[:id].to_i)
+
     if @payment
       @payment.amount = params[:amount]
       if @payment.save
@@ -61,8 +94,19 @@ class Api::PaymentsController < Api::ApiController
     end
   end
 
+
+
   def destroy
-    @payment = Payment.find_by(id: params[:id])
+    @payments = current_user.userable.payments
+    @payments_id_array = @payments.pluck(:id)
+
+    unless @payments_id_array.include?(params[:id].to_i)
+      render json: { message: "You are not authorized to view this page" } , status: :forbidden
+      return 
+    end
+
+    @payment = Payment.find_by(id: params[:id].to_i)
+
     if @payment
       if @payment.destroy
         render json: { message: "Payment destroyed" } , status: :ok
@@ -74,7 +118,13 @@ class Api::PaymentsController < Api::ApiController
     end
   end
 
-  
+
+  private
+  def is_rider?
+    unless user_signed_in? && current_user.rider?
+      render json: { message: "You are not authorized to view this page"} , status: :forbidden
+    end
+  end
