+++ app/controllers/api/vehicles_controller.rb
@@ -1,23 +1,43 @@
 class Api::VehiclesController < Api::ApiController
 
-  skip_before_action :verify_authenticity_token
+  before_action :is_driver? , except: [:get_vehicles_with_vehicle_type]
 
   def index
-    @vehicles = Vehicle.all
+
+    @vehicles = Vehicle.where(driver_no: current_user.userable.id)
     if @vehicles.empty?
       render json: { message: "No Vehicle found." } , status: :no_content
     else
       render json: @vehicles , status: :ok
     end
+
   end
 
   def show
-    @vehicle = Vehicle.find_by(id: params[:id])
-    if @vehicle
-      render json: @vehicle, status: :ok
-    else
-      render json: { message: "Vehicle not found with the id #{params[:id]}" } , status: :not_found
+    # @vehicle = Vehicle.find_by(id: params[:id])
+    # if @vehicle
+    #   render json: @vehicle, status: :ok
+    # else
+    #   render json: { message: "Vehicle not found with the id #{params[:id]}" } , status: :not_found
+    # end
+
+    @vehicles = Vehicle.where(driver_no: current_user.userable.id)
+    @vehicles_id_array = @vehicles.pluck(:id)
+
+    unless @vehicles_id_array.include?(params[:id].to_i)
+      render json: { mesage: "You are not not authorized to view this page"} , status: :forbidden
+      return 
     end
+
+
+    @vehicle = Vehicle.find_by(id: params[:id].to_i)
+
+    if @vehicle
+      render json: @vehicle , status: :ok
+    else
+      render json: { message: "No vehicle available with the id #{params[:id]}" } , status: :no_content
+    end
+
   end
 
   
@@ -28,7 +48,7 @@ class Api::VehiclesController < Api::ApiController
     @vehicle.vehicle_name = params[:name]
     @vehicle.vehicle_type = params[:type]
     @vehicle.no_of_seats = params[:seats]
-    @vehicle.driver_no = params[:driver_no]
+    @vehicle.driver_no = current_user.userable.id
 
     if @vehicle.save
       render json: @vehicle, status: :created
@@ -41,9 +61,20 @@ class Api::VehiclesController < Api::ApiController
 
   
   def update
+
+
+    @vehicles = Vehicle.where(driver_no: current_user.userable.id)
+    @vehicles_id_array = @vehicles.pluck(:id)
+
+    unless @vehicles_id_array.include?(params[:id].to_i)
+      render json: { mesage: "You are not not authorized to view this page"} , status: :forbidden
+      return 
+    end
+
+
     @vehicle = Vehicle.find_by(id: params[:id])
     if @vehicle
-      @vehicle.driver_no = params[:driver_no]
+      @vehicle.no_of_seats= params[:seats]
       if @vehicle.save
         render json: @vehicle, status: :accepted
       else
@@ -54,7 +85,19 @@ class Api::VehiclesController < Api::ApiController
     end
   end
 
+
+
   def destroy
+
+    @vehicles = Vehicle.where(driver_no: current_user.userable.id)
+    @vehicles_id_array = @vehicles.pluck(:id)
+
+    unless @vehicles_id_array.include?(params[:id].to_i)
+      render json: { mesage: "You are not not authorized to view this page"} , status: :forbidden
+      return 
+    end
+
+
     @vehicle = Vehicle.find_by(id: params[:id])
     if @vehicle
       if @vehicle.destroy
@@ -76,6 +119,16 @@ class Api::VehiclesController < Api::ApiController
     else
       render json: @vehicles , status: :ok
     end
+   end
+
+   private
+   def is_driver?
+
+    unless user_signed_in? && current_user.driver?
+      render json: { message: "You are not authorized to view this page"} , status: :forbidden
+      return 
+    end
+
    end
 
 end
